Refactor MemberProfile as a reusable component
